requirements.txt
numpy>=1.24.0
pandas>=2.0.0
geopandas>=0.13.0
shapely>=2.0.0
folium>=0.14.0
scikit-learn>=1.3.0
catboost>=1.2.0
requests>=2.31.0
tqdm
python-dotenv
streamlit
plotly
branca>=0.6.0
joblib>=1.3.0